pyproject.toml
[tool.poetry]
name = "django-relativedelta"
version = "1.1.2"
description = "Django alternative to DurationField using dateutil.relativedelta"
authors = ["Peter Bex <peter@codeyellow.nl>"]
license = "MIT"
readme = "README.rst"
repository = "https://github.com/CodeYellowBV/django-relativedelta"
homepage="https://github.com/CodeYellowBV/django-relativedelta"
keywords = ['django', 'relativedelta']
classifiers=[
        'Environment :: Web Environment',
        'Framework :: Django',
        'Operating System :: OS Independent',
        'Framework :: Django :: 2.0',
        'Framework :: Django :: 2.1',
        'Framework :: Django :: 2.2',
        'Framework :: Django :: 3.0',
        'Programming Language :: Python :: 3.5',
        'Programming Language :: Python :: 3.6',
        'Programming Language :: Python :: 3.7',
        'Programming Language :: Python :: 3.8',
        'Intended Audience :: Developers']

packages = [
    { include="relativedeltafield", from="." },
]

[tool.poetry.dependencies]
python = "^3.6"
python-dateutil = "^2.6"

[tool.poetry.dev-dependencies]
flake8 = "^3.7.9"
isort = "^4.3.21"
pytest = "^6.0.2"
poetry = "^1.0.0"
django = "^3.0.2"
tox = "^3.14.3"
django_dynamic_fixture = "^3.0.1"
pytest-coverage = "^0.0"
pytest-echo = "^1.7.1"
pytest-pythonpath = "^0.7.3"
pytest-django = "^3.7.0"
django-webtest = "^1.9.7"
psycopg2-binary = "^2.8.6"
twine = "^3.2.0"
tox-pyenv = "^1.1.0"
mysqlclient = "^2.0.1"

[tool.isort]
profile="hug"
src_paths=["isort", "test"]
combine_as_imports=true
include_trailing_comma=true
line_length=100
multi_line_output=0
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

